Enhanced Registration with ID Upload
Registration logic in the API's `RegistrationController` received a substantial update to not only include new user data fields such as `phone_num` and `id_card` but also to facilitate the uploading of ID documents during the registration process. This enhancement addresses the need for a more secure and verification-centric user onboarding experience. The inclusion of a unique phone number requirement and the ability to upload identification documents (restricted to specific formats for security and consistency) reinforces our commitment to maintaining high standards of user data integrity and security.

Additionally, the registration workflow has been wrapped in a `try-catch` block to gracefully handle any exceptions, ensuring that users receive meaningful feedback in the event of an error during registration, thus improving the overall user experience and reliability of the registration process.

This update is part of our ongoing efforts to enhance user trust and security measures on the platform.