Removed hardcoded country seeding
Streamline the seeding process by removing the hardcoded insertion of Benin into the `Country` table from `DatabaseSeeder.php`. This adjustment makes the seeders more flexible and maintainable, encouraging the use of separate seed files or external data sources for country information. This change paves the way for a more scalable and dynamic database seeding strategy, reducing potential for hard-coded data errors and improving the ease of adding or modifying country data in the future.